+++ changes.md
@@ -1,1 +1,11 @@
-v0.1.0, 23/03/2013 -- Initial release.
+**v0.1.1**, 02/04/2013
+
+**hexahedron.py**
+
+Renamed MengerSponge to Sierpinski.
+
+Added grid parameter to generate method in Sierpinski class.
+
+**v0.1.0**, 23/03/2013 
+
+Initial release.